+++ ST4R/src/api/group/getGroupDetail.js
