+++ ST4R/src/api/group/groupDelete.js
